+++ ProjectOOP/src/Process/Rule4.java
@@ -42,7 +42,7 @@ public class Rule4 extends Rules {
         Collections.sort(array,new DataSortByVolumeValue());
         for (int i=0;i<array.size();i++){
             map.put("BVname"+(i+1),array.get(i).getName());
-            map.put("BVnum"+(i+1),Integer.toString(array.get(i).getVolume()));
+            map.put("BVnum"+(i+1),Double.toString(array.get(i).getVolume()));
         }
 
         map.put("totalB",Integer.toString(filterByName(data,"^NGANHANG").getVolume()));
